+++ crates/affinidi-tdk/common/affinidi-secrets-resolver/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Affinidi Secrets Manager
 
+## 14th June 2025 (0.1.11)
+
+* **MAINTENANCE:** Updated SSI crate from 0.11 to 0.12
+* **FEATURE:** Loading JWK format keys will support public/private byte methods
+  * This has been implemented using a Inner Shadow struct for Secret `Deseralization`
+
 ## 6th June 2025 (0.1.10)
 
 * **FIX:** Generating a hash of the public key `multibase` was incorrect